backend/requirements.txt
fastapi[all]
google-auth-oauthlib
python-dotenv
SQLAlchemy
psycopg2-binary
httpx
fastapi-cors
python-dateutil
sqlalchemy-timescaledb